refactor file structure and added alerts
